requirements.txt
python-dateutil
requests
influxdb
schedule